♻️ refactor: 랭킹 시각화 방식 수정 및 리팩토링
