+++ src/app/upload/components/UploadImage.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Button } from "@/components/Button";
+import { Button } from "@/components/ui/button";
 import { XCircleIcon, PlusCircle } from "lucide-react";
 import Image from "next/image";
 import { ChangeEvent, useEffect, useState } from "react";
